+++ spine-unity/Assets/Spine/Runtime/spine-unity/Components/Following/BoundingBoxFollower.cs
@@ -49,7 +49,7 @@ namespace Spine.Unity {
 		public SkeletonRenderer skeletonRenderer;
 		[SpineSlot(dataField: "skeletonRenderer", containsBoundingBoxes: true)]
 		public string slotName;
-		public bool isTrigger;
+		public bool isTrigger, usedByEffector, usedByComposite;
 		public bool clearStateOnDisable = true;
 		#endregion
 
@@ -167,9 +167,10 @@ namespace Spine.Unity {
 						++collidersCount;
 						SkeletonUtility.SetColliderPointsLocal(bbCollider, slot, boundingBoxAttachment);
 						bbCollider.isTrigger = isTrigger;
+						bbCollider.usedByEffector = usedByEffector;
+						bbCollider.usedByComposite = usedByComposite;
 						bbCollider.enabled = false;
 						bbCollider.hideFlags = HideFlags.NotEditable;
-						bbCollider.isTrigger = IsTrigger;
 						colliderTable.Add(boundingBoxAttachment, bbCollider);
 						nameTable.Add(boundingBoxAttachment, entry.Name);
 					}